[OM-70625] set storage amount also to be a bought commodity for aurora DBs. accept 0 increment interval dependentCost. set 0 increment for aurora storage amount prices.

+++ com.vmturbo.analysis/platform.analysis/src/main/java/com/vmturbo/platform/analysis/utilities/CostFunctionFactoryHelper.java
@@ -848,15 +848,15 @@ public class CostFunctionFactoryHelper {
      * used while calculating dependentPricing.
      */
     static class SelectedStorageAndCost {
-        private final long selectedAmount;
+        private final Double selectedAmount;
         private final Double cost;
 
-        SelectedStorageAndCost(long selectedAmount, Double cost) {
+        SelectedStorageAndCost(Double selectedAmount, Double cost) {
             this.selectedAmount = selectedAmount;
             this.cost = cost;
         }
 
-        long getSelectedAmount() {
+        Double getSelectedAmount() {
             return selectedAmount;
         }
